Fixed file structure and made packages for every screen
* Separated all the screens according to figma.
