+++ androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/MainActivity.kt
@@ -6,7 +6,6 @@ import androidx.activity.compose.setContent
 import androidx.compose.material.Text
 import androidx.compose.material.Button
 import androidx.compose.runtime.Composable
-import com.EENX15_22_17.digital_journal.android.ui.arrivalpage.ArrivalPage
 import com.EENX15_22_17.digital_journal.android.ui.theme.DigitalJournalTheme
 
 
@@ -21,18 +20,6 @@ class MainActivity : ComponentActivity() {
         }
     }
 }
-
-@Composable
-fun ArrivalScreen(
-    visitId: String,
-    navBack: () -> Unit
-) {
-    ArrivalPage(
-        visitId = visitId,
-        navBack = navBack
-    )
-}
-
 
 @Composable
 fun PatientOverviewPage(
