--- androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/OverviewScreenProvider.kt
@@ -1,7 +1,0 @@
-package com.EENX15_22_17.digital_journal.android
-
-import androidx.compose.ui.tooling.preview.PreviewParameterProvider
-
-class OverviewScreenProvider : PreviewParameterProvider<()->Unit> {
-    override val values: Sequence<() -> Unit> = sequenceOf({})
-}
