+++ androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/screens/arrival/hazard/HazardAssessmentScreen.kt
@@ -1,0 +1,9 @@
+package com.EENX15_22_17.digital_journal.android.screens.arrival.hazard
+
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun TempHazardAssessment(){
+    Text(text = "HazardAssessment")
+}
