--- androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/screens/arrival/patientInfo/PatientInfo.kt
@@ -1,9 +1,0 @@
-package com.EENX15_22_17.digital_journal.android.screens.arrival.patientInfo
-
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun TempPatientInfo(){
-    Text(text = "Patientinfo")
-}
