+++ androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/screens/arrival/patientinfo/ArrivalViewModel.kt
@@ -1,4 +1,4 @@
-package com.EENX15_22_17.digital_journal.android.ui.arrivalpage
+package com.EENX15_22_17.digital_journal.android.screens.arrival.patientinfo
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
