+++ androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/screens/current/currentpatients/CurrentPatientContent.kt
@@ -3,6 +3,8 @@ package com.EENX15_22_17.digital_journal.android.screens.current.currentpatients
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.*
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.Info
@@ -12,6 +14,8 @@ import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
@@ -136,6 +140,41 @@ fun CommentsContent(visitId: String) {
                 textAlign = TextAlign.Center,
                 style = MaterialTheme.typography.h4
             )
+        }
+    }
+}
+
+// Temporary data before backend is implemented
+//TODO: Fetch data from API instead of using sampledata from the preview-provider
+private val defaultPatients: List<CurrentPatientsData> =
+    SampleCurrentPatientsProvider().values.toList()[0]
+
+@Preview
+@Composable
+fun PatientsList(
+    modifier: Modifier = Modifier,
+    @PreviewParameter(SampleCurrentPatientsProvider::class)
+    patients: List<CurrentPatientsData> = defaultPatients,
+    navigateSpecificPatient: (visitId: String) -> Unit = {},
+    navigateSpecificOverviewPage: (visitId: String) -> Unit = {}
+) {
+    val scrollState = rememberScrollState()
+    Box(
+        modifier
+            .height(400.dp)
+    ) {
+        Column(
+            Modifier.verticalScroll(scrollState)
+        ) {
+            for (patient in patients) {
+                PatientCard(
+                    visitId = patient.id,
+                    name = patient.name,
+                    securityNumber = patient.securityNumber,
+                    showOverview = { navigateSpecificOverviewPage(it) },
+                    navigateSelectedPatient = { navigateSpecificPatient(it) }
+                )
+            }
         }
     }
 }
