+++ androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/screens/current/currentpatients/CurrentPatientsScreen.kt
