+++ androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/screens/triage/contactreason/RettsClickableSection.kt
@@ -1,25 +1,18 @@
-package com.EENX15_22_17.digital_journal.android.ui.contactcause
+package com.EENX15_22_17.digital_journal.android.screens.triage.contactreason
 
 
 import androidx.compose.foundation.*
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
 import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.R
 import androidx.compose.material.icons.rounded.Close
-import androidx.compose.material.icons.rounded.Notifications
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.lifecycle.viewmodel.CreationExtras
 import com.EENX15_22_17.digital_journal.android.ui.theme.colorIcon
 import com.EENX15_22_17.digital_journal.android.ui.theme.colorTextGray
