+++ androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/screens/triage/currentHealth/CurrentHealthScreen.kt
