--- androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/screens/triage/hazardAssessment/HazardAssesment.kt
@@ -1,9 +1,0 @@
-package com.EENX15_22_17.digital_journal.android.screens.triage.hazardAssessment
-
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun TempContactReason(){
-    Text(text = "Hazardassesment")
-}
