--- androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/screens/triage/healthHistory/HealthHistory.kt
@@ -1,9 +1,0 @@
-package com.EENX15_22_17.digital_journal.android.screens.triage.healthHistory
-
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun TempHealthHistory(){
-    Text(text = "Healthhistory")
-}
