+++ androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/screens/triage/healthHistory/HealthHistoryScreen.kt
@@ -1,4 +1,4 @@
-package com.EENX15_22_17.digital_journal.android.ui.triage.history
+package com.EENX15_22_17.digital_journal.android.screens.triage.healthHistory
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.border
