+++ androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/screens/triage/healthHistory/HealthHistoryViewModel.kt
@@ -1,4 +1,4 @@
-package com.EENX15_22_17.digital_journal.android.ui.triage.history
+package com.EENX15_22_17.digital_journal.android.screens.triage.healthHistory
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
